Provide default CodeRay stylesheet if none provided by current theme.
Added ability to automagically attach a default_codemacro.css stylesheet to the current page if no .CodeRay selectors are found in any of the page's stylesheets after all page assets have been loaded.  The default_codemacro.css stylesheet is located in the plugin assets folder and copied to /public/plugin_assets at rails startup.  Similarly, there is also a new codemacro.js file that gets copied which is dynamically attached to the page via a few lines of code that appears at the end of each code listing.