Avoid CoreImage failure on attachment_fu uploads
This initializer avoids failures wen uploading images via attachment_fu
when there is no external image library to use other than CoreImage.
This works by removing "CoreImage" from the list of default processors
that attachment_fu iterates over. This avoidance is only required when
the rails process is ran under a forked process such as that which
occurs with apache + Passenger + CoreImage under OSX since the Core
Image APIs are not safe to call from a forked process and will fail. A
helpful message is written to the production.log during startup to
indicate that image resizing and thumbnailing will not be performed
unless one of the other image libs is available.